Restructure, with  own CMake for the bindings
