+++ cutensor_bindings/include/attributes.h
@@ -1,0 +1,9 @@
+#ifndef TAPP_REF_IMPL_CUTENSOR_BINDINGS_ATTRIBUTES_H_
+#define TAPP_REF_IMPL_CUTENSOR_BINDINGS_ATTRIBUTES_H_
+
+#include <tapp/attributes.h>
+#include "handle.h"
+
+#define ATTR_KEY_USE_DEVICE_MEMORY 0
+
+#endif /* TAPP_REF_IMPL_CUTENSOR_BINDINGS_ATTRIBUTES_H_ */
