+++ cutensor_bindings/include/executor.h
@@ -1,0 +1,8 @@
+#ifndef TAPP_REF_IMPL_CUTENSOR_BINDINGS_EXECUTOR_H_
+#define TAPP_REF_IMPL_CUTENSOR_BINDINGS_EXECUTOR_H_
+
+#include <tapp/executor.h>
+
+#include "error.h"
+
+#endif /* TAPP_REF_IMPL_CUTENSOR_BINDINGS_EXECUTOR_H_ */
