+++ cutensor_bindings/src/attributes.cu
@@ -1,4 +1,4 @@
-#include "cutensor_bind.h"
+#include "../include/attributes.h"
 
 TAPP_error TAPP_attr_set(TAPP_attr attr, TAPP_key key, void* value)
 {
