+++ cutensor_bindings/src/handle.cu
@@ -1,4 +1,4 @@
-#include "cutensor_bind.h"
+#include "../include/handle.h"
 
 TAPP_error TAPP_create_handle(TAPP_handle* handle)
 {
